Created KnownDevice and ConnectedDevice structs
