+++ Cargo.toml
@@ -35,3 +35,4 @@ blurz = "0.4.0"
 num_enum = "0.5.0"
 x509-parser = "0.8.0-beta3"
 hex = "0.4.2"
+mockall = "0.7.1"
